+++ packages/core/src/data-propagation/direct/DirectStreamPropagation.ts
@@ -8,11 +8,11 @@ import { DirectPropagationInterface } from './DirectPropagationInterface';
 export class DirectStreamPropagation implements DirectPropagationInterface {
   private readonly node: Libp2p;
 
-  /** Procotol used for Streaming */
-  private readonly protocol: string;
-
   /** Handler function that gets executed once message is received */
-  private handler?: (message: PropagatedMessage<any>, ctx: PropagationContext) => void;
+  private protocolHandlers: Map<
+    string,
+    (message: PropagatedMessage<any>, ctx: PropagationContext) => Promise<void> | void
+  >;
 
   /** Maximum message length that can be read */
   private readonly maxMessageBytes: number;
@@ -20,47 +20,51 @@ export class DirectStreamPropagation implements DirectPropagationInterface {
   /** Flag to stop the protocol handling */
   private stopped = false;
 
-  constructor(node: Libp2p, protocol: string, maxMessageBytes: number = 256 * 1024) {
+  constructor(node: Libp2p, maxMessageBytes: number = 256 * 1024) {
     this.node = node;
-    this.protocol = protocol;
     this.maxMessageBytes = maxMessageBytes;
-
-    this.node.handle(protocol, this.handleIncomingStream.bind(this));
+    this.protocolHandlers = new Map();
   }
 
-  private async handleIncomingStream<T>({ stream, connection }: IncomingStreamData) {
-    if (this.stopped) return;
-    try {
-      await readMessagesFromStream(
-        stream,
-        (message) => {
-          const receivedMessage = message as PropagatedMessage<T>;
-          if (this.handler) this.handler(receivedMessage, { from: connection.remotePeer, receivedAt: now() });
-        },
-        this.maxMessageBytes,
-      );
-    } finally {
-      stream.close();
-    }
-  }
-
-  async send<T>(peerId: PeerId | string, message: PropagatedMessage<T>): Promise<void> {
+  async send<T>(peerId: PeerId | string, protocol: string, message: PropagatedMessage<T>): Promise<void> {
     if (this.stopped) {
       throw new Error('DirectStreamPropagation is stopped');
     }
     const receiverPeerId = typeof peerId === 'string' ? peerIdFromString(peerId) : peerId;
-    const stream = await this.node.dialProtocol(receiverPeerId, this.protocol);
+    const stream = await this.node.dialProtocol(receiverPeerId, protocol);
     await writeToStream(stream, message);
   }
 
-  onReceive<T>(handler: (message: PropagatedMessage<T>, ctx: PropagationContext) => void): void {
-    this.handler = handler;
+  onReceive<T>(protocol: string, handler: (message: PropagatedMessage<T>, ctx: PropagationContext) => void): void {
+    this.protocolHandlers.set(protocol, handler);
+    this.node.handle(protocol, async ({ stream, connection }: IncomingStreamData) => {
+      if (this.stopped) return;
+      try {
+        await readMessagesFromStream(
+          stream,
+          (message) => {
+            const receivedMessage = message as PropagatedMessage<T>;
+            handler(receivedMessage, { from: connection.remotePeer, receivedAt: now() });
+          },
+          this.maxMessageBytes,
+        );
+      } finally {
+        stream.close();
+      }
+    });
+  }
+
+  unregisterProtocol(protocol: string): void {
+    this.protocolHandlers.delete(protocol);
+    this.node.unhandle(protocol);
   }
 
   async stop(): Promise<void> {
     if (this.stopped) return;
     this.stopped = true;
-    this.handler = undefined;
-    this.node.unhandle(this.protocol);
+    await Promise.all(Array.from(this.protocolHandlers.keys()).map((protocol) => this.node.unhandle(protocol))).catch(
+      (err) => console.error('Error stopping DirectStreamPropagation:', err),
+    );
+    this.protocolHandlers.clear();
   }
 }